version 1.2.11 fixed a potential compatibility issue re timezones.
git-svn-id: https://plugins.svn.wordpress.org/entrywizard/trunk@1068988 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ classes/ewz-base.php
@@ -33,7 +33,9 @@ abstract class Ewz_Base
         global $wpdb;
         $options = $wpdb->get_col( 'SELECT option_name FROM ' . $wpdb->options . ' WHERE option_name like "' . $prefix . '%"' );
         foreach ( $options as $option_name ) {
-            delete_option( $option_name );
+            if( preg_match( '/^ewz_[0-9]+_/', $option_name ) ){
+                delete_option( $option_name );
+            }
         }
     }
